[squash] Load CSS in stories
